+++ android/src/main/java/dev/joshi/raw_gnss/RawGnssPlugin.java
@@ -8,6 +8,7 @@ import android.location.GnssMeasurementsEvent;
 import android.location.GnssNavigationMessage;
 import android.location.LocationManager;
 import android.location.OnNmeaMessageListener;
+import android.location.GnssStatus;
 import android.os.Build;
 
 import androidx.annotation.NonNull;
@@ -28,9 +29,11 @@ public class RawGnssPlugin implements FlutterPlugin {
   private static final String GNSS_MEASUREMENT_CHANNEL_NAME =
           "dev.joshi.raw_gnss/gnss_measurement";
   private static final String GNSS_NAVIGATION_MESSAGE_CHANNEL_NAME = "dev.joshi.raw_gnss/gnss_navigation_message";
+  private static final String GNSS_STATUS_CHANNEL_NAME = "dev.joshi.raw_gnss/gnss_status";
 
   private EventChannel gnssMeasurementChannel;
   private EventChannel gnssNavigationMessageChannel;
+  private EventChannel gnssStatusChannel;
   private LocationManager locationManager;
   private Context context;
 
@@ -66,7 +69,7 @@ public class RawGnssPlugin implements FlutterPlugin {
     locationManager = (LocationManager) context.getSystemService(Context.LOCATION_SERVICE);
     gnssMeasurementChannel = new EventChannel(messenger, GNSS_MEASUREMENT_CHANNEL_NAME);
     gnssNavigationMessageChannel = new EventChannel(messenger, GNSS_NAVIGATION_MESSAGE_CHANNEL_NAME);
-
+    gnssStatusChannel = new EventChannel(messenger, GNSS_STATUS_CHANNEL_NAME);
 
     final GnssMeasurementHandlerImpl gnssMeasurementStreamHandler =
             new GnssMeasurementHandlerImpl(locationManager);
@@ -76,10 +79,14 @@ public class RawGnssPlugin implements FlutterPlugin {
             new GnssNavigationMessageHandlerImpl(locationManager);
     gnssNavigationMessageChannel.setStreamHandler(gnssNavigationMessageHandler);
 
+    final GnssStatusHandlerImpl gnssStatusStreamHandler =
+            new GnssStatusHandlerImpl(locationManager);
+    gnssStatusChannel.setStreamHandler(gnssStatusStreamHandler);
   }
 
   private void teardownEventChannels() {
     gnssMeasurementChannel.setStreamHandler(null);
     gnssNavigationMessageChannel.setStreamHandler(null);
+    gnssStatusChannel.setStreamHandler(null);
   }
 }