+++ lib/raw_gnss.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:flutter/services.dart';
 import 'package:raw_gnss/gnss_measurement_model.dart';
+import 'package:raw_gnss/gnss_status_model.dart';
 
 class RawGnss {
   /// This channel hooks onto the stream for GnssMeasurement events
@@ -12,8 +13,13 @@ class RawGnss {
   static const EventChannel _gnssNavigationMessageEventChannel =
       EventChannel('dev.joshi.raw_gnss/gnss_navigation_message');
 
+  /// This channel hooks onto the stream for GnssNavigationMessage events
+  static const EventChannel _gnssStatusEventChannel =
+      EventChannel('dev.joshi.raw_gnss/gnss_status');
+
   Stream<GnssMeasurementModel>? _gnssMeasurementEvents;
   Stream? _gnssNavigationMessageEvents;
+  Stream<GnssStatusModel>? _gnssStatusEvents;
 
   /// Getter for GnssMeasurement events
   Stream<GnssMeasurementModel> get gnssMeasurementEvents {
@@ -33,5 +39,15 @@ class RawGnss {
           _gnssNavigationMessageEventChannel.receiveBroadcastStream();
     }
     return _gnssNavigationMessageEvents!;
+  }
+
+  /// Getter for GnssMeasurement events
+  Stream<GnssStatusModel> get gnssStatusEvents {
+    if (_gnssMeasurementEvents == null) {
+      _gnssStatusEvents = _gnssStatusEventChannel.receiveBroadcastStream().map(
+          (event) => GnssStatusModel.fromJson(
+              (event as Map<dynamic, dynamic>).cast()));
+    }
+    return _gnssStatusEvents!;
   }
 }