+++ CHANGELOG.md
@@ -1,4 +1,10 @@
+## 1.1.0-beta.0
+
+* Added GNSS Status Model class
+* GNSS Status stream now provided respective class model
+
 ## 1.0.0
+
 * Upgrade to null safety
 * All model properties changed to `final`